# Split FP8 Grouped Gemm into dynamic and static version