--- a/fbgemm_gpu/experimental/gen_ai/src/quantize/ck_extensions/fp8_rowwise_grouped/fp8_rowwise_grouped_gemm.hip
+++ b/fbgemm_gpu/experimental/gen_ai/src/quantize/ck_extensions/fp8_rowwise_grouped/fp8_rowwise_grouped_gemm.hip
@@ -353,7 +353,6 @@ std::vector<at::Tensor> f8f8bf16_rowwise_grouped(
     at::TensorList WQ,
     at::TensorList x_scale,
     at::TensorList w_scale,
-    std::optional<at::Tensor> zero_start_index_M = std::nullopt,
     std::optional<std::vector<at::Tensor>> output = std::nullopt,
     std::optional<std::string> kernel_name = std::nullopt) {
   // Check that input datatypes are valid.
@@ -407,30 +406,16 @@ std::vector<at::Tensor> f8f8bf16_rowwise_grouped(
           Y[i].dtype() == at::kBFloat16, "Output dtype must be bfloat16.");
     }
   } else {
-    // Two modes for allocating output. When m_values is provided, we need
-    // the output tensor to be contiguous and can assume M, N, and K are the
-    // same across groups. Otherwise, we can allocate each output separately.
-    if (zero_start_index_M.has_value()) {
-      int M = XQ[0].size(0);
-      int N = WQ[0].size(0);
-      // Allocate an empty output array. We will set its values to zero as part
-      // of kernel setup.
-      at::Tensor Y_full =
-          at::empty({group_count, M, N}, XQ[0].options().dtype(at::kBFloat16));
-      // Split the output into groups.
-      Y = at::unbind(Y_full, 0);
-    } else {
-      for (int i = 0; i < group_count; i++) {
-        int M = XQ[i].size(0);
-        int N = WQ[i].size(0);
-        Y.push_back(at::empty({M, N}, XQ[i].options().dtype(at::kBFloat16)));
-      }
+    for (int i = 0; i < group_count; i++) {
+      int M = XQ[i].size(0);
+      int N = WQ[i].size(0);
+      Y.push_back(at::empty({M, N}, XQ[i].options().dtype(at::kBFloat16)));
     }
   }
 
   // Prepare kernel arguments by copying them to the proper device location.
   at::Tensor kernel_args =
-      get_grouped_kernel_args(XQ, WQ, x_scale, w_scale, zero_start_index_M, Y);
+      get_grouped_kernel_args(XQ, WQ, x_scale, w_scale, std::nullopt, Y);
 
   // If provided a specific kernel implementation, dispatch to it.
   if (kernel_name.has_value()) {
@@ -455,6 +440,88 @@ std::vector<at::Tensor> f8f8bf16_rowwise_grouped(
   RowwiseGroupedKernel selected_kernel =
       rowwise_grouped_heuristic_dispatch(MaxM, MaxN, MaxK);
   return selected_kernel(XQ, WQ, x_scale, w_scale, kernel_args, Y);
+}
+
+at::Tensor f8f8bf16_rowwise_grouped_dynamic(
+    at::TensorList XQ,
+    at::TensorList WQ,
+    at::TensorList x_scale,
+    at::TensorList w_scale,
+    at::Tensor zero_start_index_M,
+    std::optional<std::string> kernel_name = std::nullopt) {
+  // Check that input datatypes are valid.
+  // First confirm that there are the same number of groups in all inputs.
+  TORCH_CHECK(
+      XQ.size() == WQ.size() && XQ.size() == x_scale.size() &&
+          XQ.size() == w_scale.size(),
+      "All inputs must have the same number of groups.");
+  int group_count = XQ.size();
+  // Iterate over inputs and check they are valid.
+  for (at::Tensor x : XQ) {
+    TORCH_CHECK(x.is_cuda() && x.is_contiguous());
+    TORCH_CHECK(x.dim() == 2, "Inputs must be 2D.");
+    TORCH_CHECK(
+        x.dtype() == at::kFloat8_e4m3fnuz,
+        "Inputs must be type float8_e4m3fnuz.");
+  }
+  for (at::Tensor w : WQ) {
+    TORCH_CHECK(w.is_cuda() && w.is_contiguous());
+    TORCH_CHECK(w.dim() == 2, "Inputs must be 2D.");
+    TORCH_CHECK(
+        w.dtype() == at::kFloat8_e4m3fnuz,
+        "Inputs must be type float8_e4m3fnuz.");
+    TORCH_CHECK(
+        w.size(0) >= 512 && w.size(1) >= 512,
+        "N and K must be at least 512 for grouped gemm. For smaller inputs, consider unrolling.");
+  }
+  for (at::Tensor xs : x_scale) {
+    TORCH_CHECK(xs.dtype() == at::kFloat, "Scales must be float32.");
+  }
+  for (at::Tensor ws : x_scale) {
+    TORCH_CHECK(ws.dtype() == at::kFloat, "Scales must be float32.");
+  }
+
+  // Create a single chunk of tensor but view it as a list for compatibility.
+  int M = XQ[0].size(0);
+  int N = WQ[0].size(0);
+  // Allocate an empty output array. We will set its values to zero as part
+  // of kernel setup.
+  at::Tensor Y_full =
+      at::empty({group_count, M, N}, XQ[0].options().dtype(at::kBFloat16));
+  // Split the output into groups.
+  std::vector<at::Tensor> Y = at::unbind(Y_full, 0);
+
+  // Prepare kernel arguments by copying them to the proper device location.
+  at::Tensor kernel_args =
+      get_grouped_kernel_args(XQ, WQ, x_scale, w_scale, zero_start_index_M, Y);
+
+  // If provided a specific kernel implementation, dispatch to it.
+  if (kernel_name.has_value()) {
+    auto it = kernel_name_map.find(kernel_name.value());
+    // If not found, raise an error.
+    TORCH_CHECK(
+        it != kernel_name_map.end(),
+        "Could not find kernel " + kernel_name.value());
+    // If found, always use requested kernel.
+    it->second(XQ, WQ, x_scale, w_scale, kernel_args, Y);
+    return Y_full;
+  }
+  // Otherwise, use heuristics to find the best kernel options.
+  // We use the largest of each shape for heuristics.
+  int MaxM = 0;
+  int MaxN = 0;
+  int MaxK = 0;
+  for (int i = 0; i < group_count; i++) {
+    MaxM = max(MaxM, XQ[i].size(0));
+    MaxN = max(MaxN, WQ[i].size(0));
+    MaxK = max(MaxK, XQ[i].size(1));
+  }
+  RowwiseGroupedKernel selected_kernel =
+      rowwise_grouped_heuristic_dispatch(MaxM, MaxN, MaxK);
+  // Run kernel to populate Y.
+  selected_kernel(XQ, WQ, x_scale, w_scale, kernel_args, Y);
+  // Return unified view of Y_full.
+  return Y_full;
 }
 
 std::vector<std::string> get_f8f8bf16_rowwise_grouped_kernels() {