Summary:
We found that the `torch.stack` call required to coalesce a list of output groups is actually quite costly as it includes a data copy. This diff splits our grouped gemm kernel into a static version that returns a list of groups and dynamic version that returns a single coalesced tensor. This lets us skip the copy for cuda graph use cases.

This helps a bit with performance, it may be more effective in E2E.

Before:
```
INFO:root:Decode: B=64, PROMPT_T=2048: elapsed: 39.88ms, FLOPs: 28.54TF/s, BW: 360.45GB/s, KV cache read: 5.00GB, Weights read: 9.38GB
```

After:
```
INFO:root:Decode: B=64, PROMPT_T=2048: elapsed: 39.13ms, FLOPs: 29.09TF/s, BW: 367.41GB/s, KV cache read: 5.00GB, Weights read: 9.38GB
```

Differential Revision: D67810956